ai-image-cropper/requirements.txt
flask==3.0.3
google-cloud-aiplatform==1.59.0
pillow==10.4.0
